requirements.txt
Faker==37.1.0
libnum==1.7.1
pycryptodome==3.11.0
